add blogger flow

+++ src/features/blogs/application/use-cases/delete-blog-to-blogger-use-case.ts
@@ -1,28 +1,33 @@
-import { NotFoundException } from '@nestjs/common';
+import { ForbiddenException, NotFoundException } from '@nestjs/common';
 import { BlogsRepository } from '../../repository/blogs.repository';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
-import { BlogsService } from '../blogs.service';
+import { BlogModelType } from '../../core/entity/blogs.entity';
 
 export class DeleteBlogToBloggerCommand {
-  constructor(public readonly blogID: string) {}
+  constructor(
+    public readonly bloggerId: string,
+    public readonly blogID: string,
+  ) {}
 }
 
 @CommandHandler(DeleteBlogToBloggerCommand)
 export class DeleteBlogToBloggerUseCase
   implements ICommandHandler<DeleteBlogToBloggerCommand>
 {
-  constructor(
-    protected blogRepository: BlogsRepository,
-    protected blogService: BlogsService,
-  ) {}
+  constructor(protected blogRepository: BlogsRepository) {}
 
   async execute(command: DeleteBlogToBloggerCommand) {
-    const { blogID } = command;
+    const { bloggerId, blogID } = command;
 
-    const findBlog: boolean = await this.blogService.checkBlog(blogID);
+    const findBlog: BlogModelType | null =
+      await this.blogRepository.findBlogById(blogID);
 
     if (!findBlog) {
       throw new NotFoundException('blog not found');
+    }
+
+    if (findBlog.bloggerId !== bloggerId) {
+      throw new ForbiddenException('The user is not the owner of the blog');
     }
 
     await this.blogRepository.deleteBlog(blogID);
